Fix recipe image save path and move UploadRecipeImage component
