+++ app/ui/components/images/__init__.py
@@ -1,0 +1,1 @@
+from .upload_recipe_image import UploadRecipeImage
